requirements.txt
torch
transformers
faiss-cpu
numpy
opencv-python
pillow
torchvision
llm2vec